model/requirements.txt
absl-py==2.1.0
alembic==1.13.2
aniso8601==9.0.1
astunparse==1.6.3
blinker==1.8.2
boto3==1.34.145
botocore==1.34.145
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
Deprecated==1.2.14
docker==7.1.0
entrypoints==0.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.65.1
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
ml-dtypes==0.4.0
mlflow==2.14.3
namex==0.0.8
numpy==1.26.4
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.3
pyarrow==15.0.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.31
sqlparse==0.5.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
